requirement.txt
systemctl enable tor # enable tor every boot

status 204 no content character verification

amazon.com/dp/asin ürünlerin sayfasına gidiş için.

sayfa sayıları yanlış verilebilir check

thread yönetimi.

ürün + detay tüm varyantları bul.

consumer içerisinde producer çalışıyor.

target location check sürekli

target üzerinde ürün aramasını yap ykarıdaki tüm detayları tarat

post code girme element üzerinde


produce et  consumerlar asin al kotrol et-dataları çek

status lar 404 not shipped unavailable

varyantları listeye eklerken kontrol et